+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @loggerjs/browser
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @loggerjs/core@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes